pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gentooui"
version = "0.1.0"
description = "A Text User Interface for Gentoo Linux Installation"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Your Name", email = "your.email@example.com"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.8"
dependencies = [
    "textual>=0.41.0",
    "rich>=13.5.0",
    "psutil>=5.9.0", 
    "pexpect>=4.8.0",
    "PyYAML>=6.0.1",
    "Jinja2>=3.1.2",
    "marshmallow>=3.20.0",
    "click>=8.1.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
docs = [
    "Sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.scripts]
gentooui = "gentooui.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gentooui = ["configs/*.yaml", "templates/*.j2"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
